pyproject.toml
[tool.poetry]
name = "sso-switchover-agent"
version = "0.1.1"
description = ""
authors = ["SSO Team"]
readme = "README.md"
packages = [{include = "sso_switchover_agent"}]

[tool.poetry.dependencies]
python = "3.12.11"
fastapi = "0.115.0"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
requests = "^2.32.3"
uvicorn = "^0.30.6"
python-hosts = "^1.0.7"
pytz = "^2024.2"
pygithub = "^2.6.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[tool.autopep8]
max_line_length = 120
ignore = ["E501", "W6"]
in-place = true
recursive = true
aggressive = 3

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
